+++ js/admin-settings.js
@@ -26,4 +26,23 @@ https://ssl.webaware.net.au/
 		}
 	}
 
+	$.ajax({
+		url:		sslfix.ajax_url_wp,
+		data:		{ action: "sslfix-test-https" },
+		dataType:	"json",
+		method:		"GET",
+		xhrFields:	{ withCredentials: true },
+		success:	showHttpsDetected
+	});
+
+	/**
+	* show whether HTTPS was detected correctly within WordPress
+	* @param {Object} response
+	*/
+	function showHttpsDetected(response) {
+		if (response.https) {
+			$("#sslfix-https-detection").addClass("dashicons dashicons-" + response.https);
+		}
+	}
+
 })(jQuery);