+++ changelog.md
@@ -7,6 +7,7 @@
 * added: French translation (thanks, Houzepha Taheraly!)
 * added: can define `SSLFIX_PLUGIN_NO_HTTPS_DETECT` in wp-config.php to prevent the proxy fix, e.g. to overcome plugin conflicts
 * added: fix inline CSS background image rules, e.g. in Capture level
+* added: indicate whether WordPress HTTPS detection is successful with tick/cross on Settings page
 
 ### 2.1.3, 2015-10-05